refactor: split organism types
