--- packages/types/src/page/organisms.d.ts
@@ -1,2196 +1,0 @@
-import { z } from "zod";
-import { type PageComponentBase } from "./base";
-export interface HeroBannerComponent extends PageComponentBase {
-    type: "HeroBanner";
-    slides?: {
-        src: string;
-        alt?: string;
-        headlineKey: string;
-        ctaKey: string;
-    }[];
-}
-export declare const heroBannerComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"HeroBanner">;
-    slides: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        src: z.ZodString;
-        alt: z.ZodOptional<z.ZodString>;
-        headlineKey: z.ZodString;
-        ctaKey: z.ZodString;
-    }, "strip", z.ZodTypeAny, {
-        src: string;
-        headlineKey: string;
-        ctaKey: string;
-        alt?: string | undefined;
-    }, {
-        src: string;
-        headlineKey: string;
-        ctaKey: string;
-        alt?: string | undefined;
-    }>, "many">>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"HeroBanner">;
-    slides: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        src: z.ZodString;
-        alt: z.ZodOptional<z.ZodString>;
-        headlineKey: z.ZodString;
-        ctaKey: z.ZodString;
-    }, "strip", z.ZodTypeAny, {
-        src: string;
-        headlineKey: string;
-        ctaKey: string;
-        alt?: string | undefined;
-    }, {
-        src: string;
-        headlineKey: string;
-        ctaKey: string;
-        alt?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"HeroBanner">;
-    slides: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        src: z.ZodString;
-        alt: z.ZodOptional<z.ZodString>;
-        headlineKey: z.ZodString;
-        ctaKey: z.ZodString;
-    }, "strip", z.ZodTypeAny, {
-        src: string;
-        headlineKey: string;
-        ctaKey: string;
-        alt?: string | undefined;
-    }, {
-        src: string;
-        headlineKey: string;
-        ctaKey: string;
-        alt?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface ProductGridComponent extends PageComponentBase {
-    type: "ProductGrid";
-    skus?: string[];
-    collectionId?: string;
-    mode?: "collection" | "manual";
-    /** Enable product quick view modal */
-    quickView?: boolean;
-}
-export declare const productGridComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ProductGrid">;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ProductGrid">;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ProductGrid">;
-}, z.ZodTypeAny, "passthrough">>;
-export interface ProductCarouselComponent extends PageComponentBase {
-    type: "ProductCarousel";
-    skus?: string[];
-    collectionId?: string;
-    mode?: "collection" | "manual";
-    /** Enable product quick view modal */
-    quickView?: boolean;
-}
-export declare const productCarouselComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ProductCarousel">;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ProductCarousel">;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ProductCarousel">;
-}, z.ZodTypeAny, "passthrough">>;
-export interface RecommendationCarouselComponent extends PageComponentBase {
-    type: "RecommendationCarousel";
-    endpoint: string;
-}
-export declare const recommendationCarouselComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"RecommendationCarousel">;
-    endpoint: z.ZodString;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"RecommendationCarousel">;
-    endpoint: z.ZodString;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"RecommendationCarousel">;
-    endpoint: z.ZodString;
-}, z.ZodTypeAny, "passthrough">>;
-export interface GalleryComponent extends PageComponentBase {
-    type: "Gallery";
-    images?: {
-        src: string;
-        alt?: string;
-    }[];
-}
-export declare const galleryComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"Gallery">;
-    images: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        src: z.ZodString;
-        alt: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        src: string;
-        alt?: string | undefined;
-    }, {
-        src: string;
-        alt?: string | undefined;
-    }>, "many">>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"Gallery">;
-    images: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        src: z.ZodString;
-        alt: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        src: string;
-        alt?: string | undefined;
-    }, {
-        src: string;
-        alt?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"Gallery">;
-    images: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        src: z.ZodString;
-        alt: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        src: string;
-        alt?: string | undefined;
-    }, {
-        src: string;
-        alt?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface LookbookComponent extends PageComponentBase {
-    type: "Lookbook";
-    src?: string;
-    alt?: string;
-    hotspots?: {
-        sku?: string;
-        x: number;
-        y: number;
-    }[];
-}
-export declare const lookbookComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"Lookbook">;
-    src: z.ZodOptional<z.ZodString>;
-    alt: z.ZodOptional<z.ZodString>;
-    hotspots: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        sku: z.ZodOptional<z.ZodString>;
-        x: z.ZodNumber;
-        y: z.ZodNumber;
-    }, "strip", z.ZodTypeAny, {
-        x: number;
-        y: number;
-        sku?: string | undefined;
-    }, {
-        x: number;
-        y: number;
-        sku?: string | undefined;
-    }>, "many">>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"Lookbook">;
-    src: z.ZodOptional<z.ZodString>;
-    alt: z.ZodOptional<z.ZodString>;
-    hotspots: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        sku: z.ZodOptional<z.ZodString>;
-        x: z.ZodNumber;
-        y: z.ZodNumber;
-    }, "strip", z.ZodTypeAny, {
-        x: number;
-        y: number;
-        sku?: string | undefined;
-    }, {
-        x: number;
-        y: number;
-        sku?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"Lookbook">;
-    src: z.ZodOptional<z.ZodString>;
-    alt: z.ZodOptional<z.ZodString>;
-    hotspots: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        sku: z.ZodOptional<z.ZodString>;
-        x: z.ZodNumber;
-        y: z.ZodNumber;
-    }, "strip", z.ZodTypeAny, {
-        x: number;
-        y: number;
-        sku?: string | undefined;
-    }, {
-        x: number;
-        y: number;
-        sku?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface ImageSliderComponent extends PageComponentBase {
-    type: "ImageSlider";
-    slides?: {
-        src: string;
-        alt?: string;
-        caption?: string;
-    }[];
-    minItems?: number;
-    maxItems?: number;
-}
-export declare const imageSliderComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ImageSlider">;
-    slides: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        src: z.ZodString;
-        alt: z.ZodOptional<z.ZodString>;
-        caption: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        src: string;
-        alt?: string | undefined;
-        caption?: string | undefined;
-    }, {
-        src: string;
-        alt?: string | undefined;
-        caption?: string | undefined;
-    }>, "many">>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ImageSlider">;
-    slides: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        src: z.ZodString;
-        alt: z.ZodOptional<z.ZodString>;
-        caption: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        src: string;
-        alt?: string | undefined;
-        caption?: string | undefined;
-    }, {
-        src: string;
-        alt?: string | undefined;
-        caption?: string | undefined;
-    }>, "many">>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ImageSlider">;
-    slides: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        src: z.ZodString;
-        alt: z.ZodOptional<z.ZodString>;
-        caption: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        src: string;
-        alt?: string | undefined;
-        caption?: string | undefined;
-    }, {
-        src: string;
-        alt?: string | undefined;
-        caption?: string | undefined;
-    }>, "many">>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface ContactFormWithMapComponent extends PageComponentBase {
-    type: "ContactFormWithMap";
-    mapSrc?: string;
-}
-export declare const contactFormWithMapComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ContactFormWithMap">;
-    mapSrc: z.ZodOptional<z.ZodString>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ContactFormWithMap">;
-    mapSrc: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"ContactFormWithMap">;
-    mapSrc: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface StoreLocatorBlockComponent extends PageComponentBase {
-    type: "StoreLocatorBlock";
-    locations?: {
-        lat?: number;
-        lng?: number;
-        label?: string;
-    }[];
-    zoom?: number;
-}
-export declare const storeLocatorBlockComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"StoreLocatorBlock">;
-    locations: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        lat: z.ZodOptional<z.ZodNumber>;
-        lng: z.ZodOptional<z.ZodNumber>;
-        label: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        label?: string | undefined;
-        lat?: number | undefined;
-        lng?: number | undefined;
-    }, {
-        label?: string | undefined;
-        lat?: number | undefined;
-        lng?: number | undefined;
-    }>, "many">>;
-    zoom: z.ZodOptional<z.ZodNumber>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"StoreLocatorBlock">;
-    locations: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        lat: z.ZodOptional<z.ZodNumber>;
-        lng: z.ZodOptional<z.ZodNumber>;
-        label: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        label?: string | undefined;
-        lat?: number | undefined;
-        lng?: number | undefined;
-    }, {
-        label?: string | undefined;
-        lat?: number | undefined;
-        lng?: number | undefined;
-    }>, "many">>;
-    zoom: z.ZodOptional<z.ZodNumber>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"StoreLocatorBlock">;
-    locations: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        lat: z.ZodOptional<z.ZodNumber>;
-        lng: z.ZodOptional<z.ZodNumber>;
-        label: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        label?: string | undefined;
-        lat?: number | undefined;
-        lng?: number | undefined;
-    }, {
-        label?: string | undefined;
-        lat?: number | undefined;
-        lng?: number | undefined;
-    }>, "many">>;
-    zoom: z.ZodOptional<z.ZodNumber>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface BlogListingComponent extends PageComponentBase {
-    type: "BlogListing";
-    posts?: {
-        title: string;
-        excerpt?: string;
-        url?: string;
-        shopUrl?: string;
-    }[];
-}
-export declare const blogListingComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"BlogListing">;
-    posts: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        title: z.ZodString;
-        excerpt: z.ZodOptional<z.ZodString>;
-        url: z.ZodOptional<z.ZodString>;
-        shopUrl: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        title: string;
-        url?: string | undefined;
-        excerpt?: string | undefined;
-        shopUrl?: string | undefined;
-    }, {
-        title: string;
-        url?: string | undefined;
-        excerpt?: string | undefined;
-        shopUrl?: string | undefined;
-    }>, "many">>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"BlogListing">;
-    posts: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        title: z.ZodString;
-        excerpt: z.ZodOptional<z.ZodString>;
-        url: z.ZodOptional<z.ZodString>;
-        shopUrl: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        title: string;
-        url?: string | undefined;
-        excerpt?: string | undefined;
-        shopUrl?: string | undefined;
-    }, {
-        title: string;
-        url?: string | undefined;
-        excerpt?: string | undefined;
-        shopUrl?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"BlogListing">;
-    posts: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        title: z.ZodString;
-        excerpt: z.ZodOptional<z.ZodString>;
-        url: z.ZodOptional<z.ZodString>;
-        shopUrl: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        title: string;
-        url?: string | undefined;
-        excerpt?: string | undefined;
-        shopUrl?: string | undefined;
-    }, {
-        title: string;
-        url?: string | undefined;
-        excerpt?: string | undefined;
-        shopUrl?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface TestimonialsComponent extends PageComponentBase {
-    type: "Testimonials";
-    testimonials?: {
-        quote: string;
-        name?: string;
-    }[];
-}
-export declare const testimonialsComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"Testimonials">;
-    testimonials: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        quote: z.ZodString;
-        name: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        quote: string;
-        name?: string | undefined;
-    }, {
-        quote: string;
-        name?: string | undefined;
-    }>, "many">>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"Testimonials">;
-    testimonials: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        quote: z.ZodString;
-        name: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        quote: string;
-        name?: string | undefined;
-    }, {
-        quote: string;
-        name?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"Testimonials">;
-    testimonials: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        quote: z.ZodString;
-        name: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        quote: string;
-        name?: string | undefined;
-    }, {
-        quote: string;
-        name?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface PricingTableComponent extends PageComponentBase {
-    type: "PricingTable";
-    plans?: {
-        title: string;
-        price: string;
-        features: string[];
-        ctaLabel: string;
-        ctaHref: string;
-        featured?: boolean;
-    }[];
-}
-export declare const pricingTableComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"PricingTable">;
-    plans: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        title: z.ZodString;
-        price: z.ZodString;
-        features: z.ZodArray<z.ZodString, "many">;
-        ctaLabel: z.ZodString;
-        ctaHref: z.ZodString;
-        featured: z.ZodOptional<z.ZodBoolean>;
-    }, "strip", z.ZodTypeAny, {
-        title: string;
-        price: string;
-        features: string[];
-        ctaLabel: string;
-        ctaHref: string;
-        featured?: boolean | undefined;
-    }, {
-        title: string;
-        price: string;
-        features: string[];
-        ctaLabel: string;
-        ctaHref: string;
-        featured?: boolean | undefined;
-    }>, "many">>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"PricingTable">;
-    plans: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        title: z.ZodString;
-        price: z.ZodString;
-        features: z.ZodArray<z.ZodString, "many">;
-        ctaLabel: z.ZodString;
-        ctaHref: z.ZodString;
-        featured: z.ZodOptional<z.ZodBoolean>;
-    }, "strip", z.ZodTypeAny, {
-        title: string;
-        price: string;
-        features: string[];
-        ctaLabel: string;
-        ctaHref: string;
-        featured?: boolean | undefined;
-    }, {
-        title: string;
-        price: string;
-        features: string[];
-        ctaLabel: string;
-        ctaHref: string;
-        featured?: boolean | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"PricingTable">;
-    plans: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        title: z.ZodString;
-        price: z.ZodString;
-        features: z.ZodArray<z.ZodString, "many">;
-        ctaLabel: z.ZodString;
-        ctaHref: z.ZodString;
-        featured: z.ZodOptional<z.ZodBoolean>;
-    }, "strip", z.ZodTypeAny, {
-        title: string;
-        price: string;
-        features: string[];
-        ctaLabel: string;
-        ctaHref: string;
-        featured?: boolean | undefined;
-    }, {
-        title: string;
-        price: string;
-        features: string[];
-        ctaLabel: string;
-        ctaHref: string;
-        featured?: boolean | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface TestimonialSliderComponent extends PageComponentBase {
-    type: "TestimonialSlider";
-    testimonials?: {
-        quote: string;
-        name?: string;
-    }[];
-}
-export declare const testimonialSliderComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"TestimonialSlider">;
-    testimonials: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        quote: z.ZodString;
-        name: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        quote: string;
-        name?: string | undefined;
-    }, {
-        quote: string;
-        name?: string | undefined;
-    }>, "many">>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"TestimonialSlider">;
-    testimonials: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        quote: z.ZodString;
-        name: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        quote: string;
-        name?: string | undefined;
-    }, {
-        quote: string;
-        name?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"TestimonialSlider">;
-    testimonials: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        quote: z.ZodString;
-        name: z.ZodOptional<z.ZodString>;
-    }, "strip", z.ZodTypeAny, {
-        quote: string;
-        name?: string | undefined;
-    }, {
-        quote: string;
-        name?: string | undefined;
-    }>, "many">>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface GiftCardBlockComponent extends PageComponentBase {
-    type: "GiftCardBlock";
-    denominations?: number[];
-    description?: string;
-}
-export declare const giftCardBlockComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"GiftCardBlock">;
-    denominations: z.ZodOptional<z.ZodArray<z.ZodNumber, "many">>;
-    description: z.ZodOptional<z.ZodString>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"GiftCardBlock">;
-    denominations: z.ZodOptional<z.ZodArray<z.ZodNumber, "many">>;
-    description: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"GiftCardBlock">;
-    denominations: z.ZodOptional<z.ZodArray<z.ZodNumber, "many">>;
-    description: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface PopupModalComponent extends PageComponentBase {
-    type: "PopupModal";
-    trigger?: "load" | "delay" | "exit";
-    delay?: number;
-    content?: string;
-}
-export declare const popupModalComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"PopupModal">;
-    trigger: z.ZodOptional<z.ZodEnum<["load", "delay", "exit"]>>;
-    delay: z.ZodOptional<z.ZodNumber>;
-    content: z.ZodOptional<z.ZodString>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"PopupModal">;
-    trigger: z.ZodOptional<z.ZodEnum<["load", "delay", "exit"]>>;
-    delay: z.ZodOptional<z.ZodNumber>;
-    content: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-} & {
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-} & {
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"PopupModal">;
-    trigger: z.ZodOptional<z.ZodEnum<["load", "delay", "exit"]>>;
-    delay: z.ZodOptional<z.ZodNumber>;
-    content: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface CollectionListComponent extends PageComponentBase {
-    type: "CollectionList";
-    collections?: {
-        id: string;
-        title: string;
-        image: string;
-    }[];
-    desktopItems?: number;
-    tabletItems?: number;
-    mobileItems?: number;
-}
-export declare const collectionListComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"CollectionList">;
-    collections: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        id: z.ZodString;
-        title: z.ZodString;
-        image: z.ZodString;
-    }, "strip", z.ZodTypeAny, {
-        id: string;
-        title: string;
-        image: string;
-    }, {
-        id: string;
-        title: string;
-        image: string;
-    }>, "many">>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"CollectionList">;
-    collections: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        id: z.ZodString;
-        title: z.ZodString;
-        image: z.ZodString;
-    }, "strip", z.ZodTypeAny, {
-        id: string;
-        title: string;
-        image: string;
-    }, {
-        id: string;
-        title: string;
-        image: string;
-    }>, "many">>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-} & {
-    type: z.ZodLiteral<"CollectionList">;
-    collections: z.ZodOptional<z.ZodArray<z.ZodObject<{
-        id: z.ZodString;
-        title: z.ZodString;
-        image: z.ZodString;
-    }, "strip", z.ZodTypeAny, {
-        id: string;
-        title: string;
-        image: string;
-    }, {
-        id: string;
-        title: string;
-        image: string;
-    }>, "many">>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-}, z.ZodTypeAny, "passthrough">>;
-//# sourceMappingURL=organisms.d.ts.map
