+++ packages/types/src/page/organisms/index.d.ts
@@ -1,0 +1,17 @@
+export * from "./HeroBanner";
+export * from "./ProductGrid";
+export * from "./ProductCarousel";
+export * from "./RecommendationCarousel";
+export * from "./Gallery";
+export * from "./Lookbook";
+export * from "./ImageSlider";
+export * from "./ContactFormWithMap";
+export * from "./StoreLocatorBlock";
+export * from "./BlogListing";
+export * from "./Testimonials";
+export * from "./PricingTable";
+export * from "./TestimonialSlider";
+export * from "./GiftCardBlock";
+export * from "./PopupModal";
+export * from "./CollectionList";
+//# sourceMappingURL=index.d.ts.map
